added 3 restaurants
