+++ config1/config.json
@@ -3,8 +3,8 @@
   "port": 8081,
   "host": "dhall1",
   "timeunit": 100,
-  "nr_waiters": 4,
-  "nr_tables": 7,
+  "nr_waiters": 3,
+  "nr_tables": 6,
   "max_items_order": 10,
   "max_orderid": 500,
   "min_waiter_pickuptime": 2,
