+++ config3/menu.json
@@ -1,0 +1,58 @@
+[
+  {
+    "id": 1,
+    "name": "pizza",
+    "preparation-time": 20,
+    "complexity": 2,
+    "cooking-apparatus": "oven"
+  },
+  {
+    "id": 2,
+    "name": "salad",
+    "preparation-time": 10,
+    "complexity": 1,
+    "cooking-apparatus": null
+  },
+  {
+    "id": 3,
+    "name": "zeama",
+    "preparation-time": 7,
+    "complexity": 1,
+    "cooking-apparatus": "stove"
+  },
+  {
+    "id": 4,
+    "name": "Scallop Sashimi with Meyer Lemon Confit",
+    "preparation-time": 32,
+    "complexity": 3,
+    "cooking-apparatus": null
+  },
+  {
+    "id": 5,
+    "name": "Island Duck with Mulberry Mustard",
+    "preparation-time": 35,
+    "complexity": 3,
+    "cooking-apparatus": "oven"
+  },
+  {
+    "id": 6,
+    "name": "Waffles",
+    "preparation-time": 10,
+    "complexity": 1,
+    "cooking-apparatus": "stove"
+  },
+  {
+    "id": 7,
+    "name": "Aubergine",
+    "preparation-time": 20,
+    "complexity": 2,
+    "cooking-apparatus": null
+  },
+  {
+    "id": 8,
+    "name": "Lasagna",
+    "preparation-time": 30,
+    "complexity": 2,
+    "cooking-apparatus": "oven"
+  }
+]
